refactor(testdriver-api): move TestTimingContext to where it is being used
